Consolidate tokenizer unit test files
