+++ maze_dataset/maze/lattice_maze.py
@@ -566,6 +566,8 @@ class LatticeMaze(SerializableDataclass):
     ) -> "LatticeMaze":
         if isinstance(maze_tokenizer, TokenizationMode):
             maze_tokenizer = MazeTokenizer(maze_tokenizer)
+        if isinstance(maze_tokenizer, MazeTokenizer2) and maze_tokenizer not in [MazeTokenizer2.from_legacy(tm) for tm in TokenizationMode]:
+            raise NotImplementedError(f"Only exact conversions of legacy tokenizers supported, not {maze_tokenizer}.")
 
         if isinstance(tokens, str):
             tokens = tokens.split()
